Rename hex_color files
