+++ lib/widgets.dart
@@ -1,4 +1,4 @@
 library;
 
-export 'src/widgets/hex_color.dart';
+export 'src/widgets/storyblok_color.dart';
 export 'src/widgets/rich_text.dart';
